added a model for manman vs ratman
THIS IS ONLY A SAMPLE.

This is a sample combat between ManMan and RatMan.
This goes on for 10 steps. The idea is whichever character ends up with lesser health out of the two at the end of 10 steps will lose the game.
Damage and recovery are assumed to be proper numerical values which are directly subtracted and added to the H[n-1] for calculating H[n]. They are not treated as rates or something.

Not using the updateHealth funtion since it is redundant in this case as it is a function with a single line body, which is replaced by a statement in a for loop, making it simpler.

#consider changing size of damage vector from 100 to 99.